// packages/bot/src/enums/TwitchGameId.ts
export enum TwitchGameId {
  Dota2 = '29595',
  Hearthstone = '138585',
  callOfDutyModernWarfare = '512710',
  counterStrikeGlobalOffensive = '32399',
  justChatting = '509658',
  leagueOfLegends = '21779',
  minecraft = '27471',
  overwatch = '488552',
  pathOfExile = '29307',
  starcraft2 = '490422',
  terraria = '31376',
  warcraft3 = '12924',
  worldOfWarcraft = '18122',
}